backend/requirements_enhanced.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
mysqlclient==2.2.0
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
PyJWT==2.8.0

# Caching & Performance
redis==5.0.1
django-redis==5.4.0
celery==5.3.4

# NLP & AI
rasa==3.6.13
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99
langdetect==1.0.9
textblob==0.17.1

# Search
django-elasticsearch-dsl==8.0
elasticsearch==8.11.0

# Image Processing
Pillow==10.1.0
reportlab==4.0.7

# API & Documentation
drf-yasg==1.21.7
django-filter==23.5

# Notifications
celery-beat==2.5.0
django-push-notifications==3.0.2
twilio==8.10.3
sendgrid==6.11.0

# Payment
stripe==7.7.0

# Analytics
django-analytical==3.1.0

# Social Auth
social-auth-app-django==5.4.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Monitoring
sentry-sdk==1.38.0

# Rate Limiting
django-ratelimit==4.1.0
